Started work on the custom admin UI
Created an admin page and restricted access to users who either own the
Google App, or else who have an email address with a domain of
potatolondon.com or potatobristol.com.

The GAE Users api doesn't seem to register if a user logs out of their
Google account so I also added in a Logout link for the app.