First pass at allowing people to add a new post.
The basics were easy here, but I an into problems trying to get a consistent
view of the blogpost list after having added a post. Eventually resolved this
by passing the primary key of the new post back to the admin page so that
it could check that it had it. Rather clunky though and it seems as if this
should be something the cache should handle, so I may be missing something.