Modified the 'view post' view so that Markdown is converted to sanitized HTML.
Used the Python Markdown library to convert user markdown to HTML and
sanitized it with the Bleach library. For consistency it would be nice to
do the server side markdown to HTML conversion using the same markdown-js
library used on the client side when entering the post, but I've left this
as a 'nice to have' TODO for now.